add user args to main and idle
sometimes one needs to dynamically construct
GUIs at runtime. Add a userdata parameter that
can be used to pass outer state to
constructed application instances. This makes
remi a little more similar to conventional GUI
frameworks

eventually it would be good to have userdata
arguments for callbacks, but this is sufficient for now.