Fix timeout behaviour on socket read
Previously on a socket timeout (which happens if you dont interact with the website for some seconds, something around 30s for me on ubuntu) the behaviour was to just fail
Now we just try to read again.
With the previous behaviour connections were lost and undesired behaviours happened.
With this one, PLEASE TRY IT, it just keeps trying to read. In my apps behaves better.