server: automatically update html when anything changes
use an event object to signal a redraw any time a message
is recieved from the client (i.e. any websocket sendCallback
call).

This allows one to remove the update timer / update interval
completely.

Also move to a lock rather than a semaphore, as I think that
is the desired semantics