Doc: let remi package be discoverable by Sphinx
Previously building the documentation using the makefile rose warnings: it was not able to import `remi` package. The problem was that the repository root folder was not in `sys.path`, thus it was not considered when `remi.rst` tried to import `remi`.

Now all modules are correctly imported and Sphinx documentation is correctly built and populated with docstrings.